modified:   workspaces/DummyRosterTwo/DummyRoster.Common.EntityModel/Models/Form.cs

+++ workspaces/DummyRosterTwo/DummyRoster.Common.EntityModel/Models/Form.cs
@@ -17,18 +17,18 @@ namespace DummyRoster.Common.EntityModel.Models;
 public partial class Form
 {
     [Key]
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     public string? Description { get; set; }
 
     [Column(TypeName = "INT")]
-    public long? CustomerId { get; set; }
+    public int? CustomerId { get; set; }
 
     [Column(TypeName = "INT")]
-    public long? CarrierId { get; set; }
+    public int? CarrierId { get; set; }
 
     [Column(TypeName = "INT")]
-    public long? EmployeeId { get; set; }
+    public int? EmployeeId { get; set; }
 
     [Column(TypeName = "DATETIME")]
     public DateTime? FormDate { get; set; }
